Explicitly require navigation actions to end with top-level context.
<p class=note>Navigation actions - Get, Back, Forward, Refresh - always,
  unless they abort the navigation, complete with the <a>current top-level browsing context</a>
  being the <a>current browsing context</a>. Even if we navigate Back or Forward to a state
  where a different <a href=https://html.spec.whatwg.org/#browsing-context>browsing context</a>
  was the <a>current browsing context</a>.

Alternative behavior *could* be for the remote end to "remember"
what was the <a>current browsing context</a> at each of the state
of the "history". I don't support that behavior because I believe
it is more complex in implementation
([speculating] other than WebDriver, browser components don't seem
to already define the notion of the <a>current browsing context</a>
so WebDriver would have to "remember" those for itself [/speculating]).

This suggestion was prompted by the regression (in this yet unspecified behavior)
in chrome/chromedriver:
https://code.google.com/p/chromedriver/issues/detail?id=1106