Cookies: correct fields retrieved from a cookie
The previous version of "serialized cookie" used definitions from the
"receives a cookie" algorithm in RFC 6265 (cookie-attribute-list et
al.) which are only used when a new cookie is created.

Section 5.3 describes what fields the user agent is mandated to store
about a cookie, and we should reference these instead.

The introduction of the new Dependencies sub-chapter is priming Add
Cookie (and partly Delete Cookie) for some upcoming changes where it
would, I think, be tedious to keep extracting fields, especially when
the Cookies specification itself is vague on this topic.

Signed-off-by: AutomatedTester <dburns@mozilla.com>