cookies: split Get Cookie into two distinct endpoints
Get Cookie previously accepted an optional URL parameter, that if
set would return a list of a single cookie if a matching cookie was
found. This patch changes it so that there is a separate endpoint for
getting the single cookie to avoid confusion as our URL matching does
not really support optional parameters. This is not a functional change.

Additionally this patch introduces a breaking chance: Instead of returning
a list with a single cookie or an empty list, it introduces a "no such
cookie" error when no matching cookie is found.