navigation: incorporate ssl certificate handling in go command
This patch removes chapter 8 on handling untrusted SSL certificates and
incorporates it in the Go command in chapter 7.

Furthermore it introduces a "secure SSL" state that is associated with
the session.  This can be set through the acceptSslCerts capability.

The exact specifics in processing this capability needs to be ironed
out, but this is currently not possible due to the way capabilities
are processed.