Overhaul the section on processing capabilities.
There are several goals with these changes:

1.  The endpoint node should either be able to satisfy all capabilities
    requested by the local end, or fail to create a session.

2.  The local end should be able to submit multiple sets of acceptable
    capabilities with a clearly defined order of preference. This eliminates
    the need for extra, potentially expensive, new session requests when one
    set of capabilities cannot be satisfied.

3.  Permit the definition and processing of custom capabilities, either by
    the endpoint node or an intermediate node.

4.  Preserve wire compatibility with existing implementations (i.e. the changes
    do not depend on the "requiredCapabilities" and "desiredCapabilities"
    parameters).

----

When the local end truly has no preference for the features supported by a
new session, it can simply request an empty set of capabilities.

    POST /session
    {"capabilities": {}}

The local end requires a specific browser (Firefox), but has no preference
for the exact value:

    POST /session
    {"capabilities": {"alwaysMatch":{"browserName":"firefox"}}}

The local end requires Firefox on Linux, or Chrome on Windows: the order
specified dictates in which order the endpoint should try to satisfy the
request:

    POST /session
    {"capabilities": {
      "firstMatch": [
        {"browserName": "firefox", "platformName": "linux"},
        {"browserName": "chrome", "platformName": "windows"}
      ]}}

Capabilities that are always required may be specified as "alwaysMatch". These
are merged with each "firstMatch" entry, eliminating the need to specify
expensive capabilities (like an encoded profile) multiple times:

    POST /session
    {"capabilities": {
       "alwaysMatch": {
         "browserName": "firefox",
         "profile": "base64-encoded-blob"
       },
       "firstMatch": [
         {"platformName": "linux"},
         {"platformName": "windows"}
       ]}}

Version strings are complicated. The exact matching algorithms are explicitly
left for each vendor, however, implementors are encouraged to support basic
operators to define a range of versions. That is, to request a version of
Firefox older than 48, the local end would specify:

    POST /session
    {"capabilities": {
       "alwaysMatch": {
         "browserName": "firefox",
         "browserVersion": "< 48"
       }}}

If an intermediate node requires additional information (i.e. auth credentials),
it MAY define custom capabilities, but it SHOULD require this information be
passed as a top level parameter:

    POST /session
    {"username": "test-user",
     "password": "abc123",
     "capabilities": {}}

Capabilities may never be ignored. If the endpoint does not recognize a key, it
must fail to match on those capabilities:

    POST /session
    {"capabilities": {
       "alwaysMatch": {
         "browserName": "firefox",
         "unrecognized-capability": 1234
       }}}

Given that unrecognized capabilities will cause new session requests to fail,
the spec requires that intermediate nodes that use capabilities for routing
remove those capabilities before forwarding the request to the selected remote
end.