Reordered list of endpoints to match navigation
The list of endpoints had a different ordering to the navigation of the
specification. This was confusing as the navigation matched the layout
and ordering of the page, making the list of endpoints the odd one out.
This update reorders the list to be in line with navigation and the
order of the page itself.