An element reference should be a UUID
But it may not be. Implementations need only ensure that the
reference is unique within the current session in order for
the protocol to work.

This allows implementors more latitude in how to create the
reference. The original IEDriver used a serialisation of the
COM object's address, and that would meet the constraints
required here too.