Fix up links identified as being dead or incorrect
In some cases, this means that we've also had to futz with
the spec text, but the logical meaning should be identical.

The W3C link validtor does not correctly check links into
PDF documents, however loading these in Chrome show that
they are valid.

I've removed the test about showModalDialog since this
method has been removed from modern browsers and is no
longer supported.

This does not fix the broken link in the execute script
section