`Element Clear`: Apply to all resettable elements
There's no reason to discriminate. We rely on the fact that
the focusing and unfocusing steps are run in order to ensure
that the correct "onchange" event is fired. Existing
implementations based on Selenium's "atoms" don't bother
firing an `input` event, and so neither do we.

As specced, this is a slow-motion way of calling a form's
`reset` method with additional focusing steps. It fulfills
the need of users to reset form fields to their default
values.

Closes #561