Introduce the concept of an associated session
An intermediary node needs to route commands to an upstream
node rather than acting upon them directly. A later PR will
include the mechanism for doing this, but for now track the
session that's been opened upstream and ensure that it will
be closed properly at the end of the session.