Simplify the introduction of New Session command
Fixes https://github.com/w3c/webdriver/issues/826

The introduction of the New Session command lists a number of reasons
for failing to create a new session, but these are already defined in
the algorithm. This causes the overhead of keeping the introduction
in sync with any algorithm changes.

Currently, the introduction was missing:
* a definition for: provisioning of a remote end
* failing due to a user prompt