Don't error for configuration properties in new session.
Configuration properties like "timeouts" that aren't matched against
the browser were errorneously causing the new session command to
fail due to over-zealous validation. In fact since we moved validation
to an upfront step, no further validation is required at this point
and it's safe to just copy any property into the capabilities
object (and indeed later parts of the spec assume that we do this).