Re-phrase language for branching logic
In the phrase, "the first matching user prompt handler," the word
"first" may be interpreted in one of two ways:

1. There are multiple user prompt handlers from which the match may draw
2. There are multiple matching options in the list that follows

The first interpretation is incorrect because a session has exactly one
value for the user prompt handler. The second interpretation is also
incorrect because the available options are mutually exclusive (so
their sequence does not matter).

Re-phrase the language to avoid this ambiguity, borrowing verbiage from
the "close the session" algorithm as currently defined in the
specification.