Simplify definition of "readiness state"
The condition pertaining to the current user prompt in the session of an
endpoint node is superseded by the condition concerning the number of
maximum active sessions. In addition, that latter condition also applies
to intermediary nodes.

Remove the redundant condition, and generalize the resulting description
to apply to all node types.