obscured: take <select multiple> into account
As the options of a <select multiple> are painted and represented as DOM
elements, the center point would end up targetting one of the <option>
elements.  We can take this into account by using select.contains(option).

If the first pointer-interactable paint tree element is itself, this is
also passes the check, since the DOMElement.contains(otherEl) performs
an _inclusive_ descendant check.  This means element.contains(element)
is true.