Correct specification of command arguments
WebDriver commands are currently written to reference "URL variables" by
name. For example (from the "Element Send Keys" command):

>     <li><p>Let <var>element</var> be the result
>      of <a>trying</a> to <a>get a known element</a>
>      with argument <var>element id</var>.

This is in conflict with the current command invocation, where
they are specified as a collection of name-value pairs.

In addition, some commands reference an argument named "parameters" that
is expected to contain the parsed request body. For example (again from
the "Element Send Keys" command);

>     <li><p>Let <var>text</var> be the result
>      of <a>getting a property</a> called "<code>text</code>"
>      from the <var>parameters</var> argument.

However, this argument is not currently provided when the commands are
invoked.

Clarify the text at the command invocation site to expand the "URL
variables" into named arguments and to include the "parameters"
argument.