There is never a `null` `page loading strategy`
In the other places where the spec talks about
the `page loading strategy` we already state
that it defaults to `normal`, and we will have
removed `null` values while processing
capabilities.

If a user attempts to set the `page loading
strategy` to `null` via `Set Timeouts`, then
that is clearly an error.