keyboard-interactable element: omit ownerDocument
It does not make sense to talk about the target of keyboard interaction
when defining which elements are keyboard interactable.  This belongs in
the key dispatch section in the chapter on actions.

I believe it is also inferred what the default target of the interaction
should be through the pointer events standard.

Fixes: https://github.com/w3c/webdriver/issues/886