delete session: close session only if it is still active
For implementation compatibility we should not try to close the session
if the session does not exist.  This can be interpreted in different
ways because of the use of the "try" algorithm.

Instead we make the Delete Session algorithm explicitly call out that
the session must be active in order to attempt closing it.