Only wait for navigation to complete if it can.
If the current browsing context is no longer open then
we can safely assume that its navigation has completed.
This might be the case where a `click` has caused some
long-running JS to complete and then close the current
div or window.

I think that this fully addresses the problems in #871.