find: Avoid reliance on property value
Because browsing contexts may modify the built-in
`Element.prototype.getElementsByTagName` property value (or define a
distinct value for a given Element instance), invoking the function as
defined by a given node may produce unpredictable results.

Modify the text to reference the built-in function generically, avoiding
this ambiguity and increasing parity with similar algorithm steps in
related locator strategies.