find: Account for WebDriver error from locator
Each locator strategy may produce an error during execution. This
behavior is distinct from returning a WebDriver error value and is
handled by the text "if [an error] occurs," in the "find" algorithm.
However, in addition to producing an error in this way, the "xpath"
locator strategy may return a WebDriver error.

Update the "find" algorithm to handle this case by relying on the
specification's "try" construct.