Go: Remove unnecessary step
The conditions described by this step (i.e. an unspecified property or a
non-string property value) are a subset of the cases described by a
subsequent step (i.e. value that is neither an absolute URL, an absolute
URL with fragment, nor a local scheme). Because the observable behavior
is identical in both cases, the former step is not necessary.

Remove the step in order to reduce algorithm complexity.