Scope extension commands to current session
Extension commands are expected to be executed in the context of the
remote end's current session. This expectation is normatively enforced
by the remote end's processing model and further suggested by the first
example in the section titled "Protocol Extensions."

Update the extension mechanism to be phrased in terms of a URI template
that includes the current session ID (instead of a URL that does not).