Account for any session-independent commands
The "New Session" command does not specify a session identifier, nor
does it depend on the "current session" variable. This is also true of
the "Status" command, and any number of extension commands may likewise
operate outside the context of any particular session.

Skip the substeps related to the current session for any command that
does not specify a "session id" variable. Introduce a non-normative note
documenting the intention of this condition since it is partially
intended to satisfy externally-defined criteria.

Furthermore, processing such commands should not involve extracting the
"session id" value from the variable named "url variables". Re-locate
that step to the branch that concerns only those commands where it is
relevant.