Account for "Status" command in processing model
Similar to the "New Session" command, the "Status" command does not
specify a session identifier, nor does it depend on the "current
session" variable. The substeps related to those values should therefore
be skipped for the "Status" command.

Furthermore, processing either command should not involve extracting the
"session id" value from the variable named "url variables". Re-locate
that step to the branch that concerns only those commands where it is
relevant.