Correct "error" example
Because the "send an error" algorithm provides error data as an
argument to the "send a response" algorithm, the JSON body is expected
to contain a "value" attribute whose value is an object describing the
error. Update the non-normative description and example to reflect this.

The remote end processing model explicitly allows invalid session IDs to
be specified to the Delete command. Update the example to use a command
which would respond to an invalid session ID with the "invalid session
id" error.