Better define proxies
This includes a fix where we had omitted "direct" as an option
and accidentally included "noproxy" as a proxy type.

This closes #837 by ensuring that all entries in the table
refer to the valid proxy type for which they should be used.

Closes #841, since the URL spec specifies that a host can
include credentials.

Closes #913 by properly spelling out `noProxy` and when it
should be used.

Closes #949 by removing the `*ProxyPort` settings, which are
not part of the original selenium wire protocol.