window rect: introduce window state concept
When an OS window is minimised/maximized/fullscreened, there is currently
no homogenous way of telling what state it is in.

This introduces a "state" field on the window rect which returns the
window's state that is applicable for the current UA and the current
system.

The motivation behind this change is to make it easier to test the
specification, but I imagine it is also useful to users who want to
query the window's current state.