maximize window: transition to maximized state and make wm provisions
A window's maximized state is often not the same as increasing its
dimensions to the maximum allowed screen size.  We need to be explicit
that the OS window should be moved into the maximized state.

Furthermore, a certain variety of window managers does not support the
concept of window maximation at all.  In this case, it is permissable for
WebDriver to increase the window's boundaries to the maximum screen size.

In the latter case, it is impossible to tell when if the window
has reached the full dimensions of the available screen, because
the window may have decorated borders.  This causes there to be an
unknown disjunction between window.screen.availHeight/availWidth and
window.outerWidth/outerHeight.  In this case we need to permit the
WebDriver implementation to return after a timeout so that it does not
hang indefinitely.