minimize window: make it possible to restore the window
It is the intention that calling the Minimize Window command a second
time will restore it to its original dimensions on the visible screen.
It looks like an oversight that the specification does not cater for this.

This patch also makes it crystal clear that minimizing or restoring
the window must happen synchronously and not return before the
visibilitychange DOM event has fired.