invalid element state: correct description
That an element is in an invalid state does not mean it has been detached
from the document.  It usually means the element is in a state in which
it is impossible for WebDriver to interact with it, e.g. if you try to use
the high-level Element Click command to click an element that is disabled.

Fixes: https://github.com/w3c/webdriver/issues/989
Reported-by: @titusfortner