Add states to propmt handling to allow notification of handled prompts
Setting a user prompt handler accepts or dismisses a prompt with no
notification to the user that the prompt was handled. In the primary use
case of testing, this hides important information from the user, that a
prompt the test did not expect was encountered, and handled. If the user
wishes to have their test ignore this case, they should make a fully
informed and conscious decision to do so. This change adds options to the
user prompt handler to allow for the case to inform the user of the
handled propmt via an error. To maintain compatibility with previous spec
language, the added values are the ones that describe the new behavior.

Closes issue #912.