Layout should be complete before screenshots are taken.
We couch this in the language of `requestAnimationFrame` so that
we can be sure that there are no pending paint events. This does
not guarantee that all resources on the page are fully laoded.

This also brings the language used in `Take Element Screenshot`
into line with the language used in `Take Screenshot` by using
`trying` instead of checking the return value.

Starts to address #895