Acknowledge the evil-bitness of interface change
Since many of the implementations of webdriver are OSS
it should be trivial to remove the interface change and
make automation undetectable unless you use some other
means to spot it.

The comment itself is frivolous and unhelpful to spec
developers, but it's quite fun to have it be there.