Ensure deserialization of parameters happens before other checks.
For an implementation that deseralizes the parameters in a frontend
before contacting the browser it's difficult to check for unexpected
alerts or perform other actions that require the browser before
completing deserialization. So ensure that deserialization is always
fully complete before taking other actions. In practice the only
observable change here is the error observed if there is both an
invalid request and e.g. an unexpected alert open.