is stale: take shadow root into account
The current algorithm for finding out if an element is stale, that is
conencted to the current document, does not take into account the fact
that an element may have a shadowRoot element as its document element.
Additionally, this shadow root may sit inside a possibly infinite number
of othe shadow roots.

Fixes: https://github.com/w3c/webdriver/issues/1055