Simplify language when checking if session is present
The text modified is already in a block where we've checked
for the presence of the session id in the url variables. If
there's no session, we should exit immediately. Moved that
step up to bring it closer to the loading of the session and
the declaration of session id from the url variables.