is stale: take document into account
The check for whether an element is connected does not take into account
that the document may have changed.  This is a special case for WebDriver,
where element references are preserved across documents.

To work around this issue we compare the element's node document
(element.ownerDocument) to the current browsing context's active document
(element.ownerDocument.defaultView.document).