set window rect: restrict to signed 32-bit integer input
The screenX/screenY/outeWidth/outerHeight attributes on the Window
interface should according to CSSOM-VIEW be longs.  WebIDL defines longs
as platform-independent signed 32-bit integers, which means the bitness
of the platform does not come into effect.

We don't want the input to WindowProxy.moveTo or resizeTo to be larger
than the supported numbers it can take.  This patch reduces the allowed
sizes to i32s.