delete session: remove inaccurate non-normative paragraph
It does not make any sense to say that the connection is being
terminated as the wire protocol is implemented over HTTP.

It also mentions that the top-level browsing contexts associated
with the current session should be closed, which is not the case with
all drivers.

The whole paragraph, overall, is informative and could easily be
mistaken for a normative paragraph.