clear a resettable element: check for files and constraint validations
Previously Element Clear only looked at the value IDL attribute.
Following this patch implementations must also look at the selected
files array in case the element is an <input type=file> since its
value is always an empty string.

It also asks implementations to consider constraint validations.
This is in reality only a bug in Gecko where it is possible to type
illegal characters into <input type=number>.  In Blink this is not
possible and it is not susceptible to the same problem.

If we do not check the element's validity first, we run the risk
of making Element Clear a no-op because the value IDL attribute
isn't updated until the user's input is valid.  This means we return
early.