interface: expose navigator.webdriver unconditionally
The navigator.webdriver WebIDL attribute is currently only meant
to be exposed when the webdriver-active flag is true.  Safari and
Edge expose the attribute unconditionally, and Gecko plans to follow
suit.

Making the attribute selectively present, which is unlike any other
web features are implemented, forces web authors to detect automation
in the following way:

	var underAutomation = ("webdriver" in navigator) ? navigator.webdriver : false;

Following this patch, web authors can do this:

	var underAutomation = navigator.webdriver;

Fixes: https://github.com/w3c/webdriver/issues/1214