take element screenshot: drop scroll feature
Issue https://github.com/w3c/webdriver/issues/1245 points out that
you cannot pass data parameters to a GET method request, so it is
currently impossible to pass a "scroll" field to Take Element
Screenshot.

It was also discovered in the discussion that no vendors implement
this feature.  Because we want the specification to match reality,
this patch removes the "scroll" field for Take Element Screenshot.

Fixes: https://github.com/w3c/webdriver/issues/1245