take element screenshot: drop scroll feature
Issue https://github.com/w3c/webdriver/issues/1245 points out that
you cannot pass data parameters to a GET method request, so it is
currently impossible to pass a "scroll" field to Take Element
Screenshot.

It was also discovered in the discussion that no vendors implement
this feature.  Because we want the specification to match reality,
this patch removes the "scroll" field for Take Element Screenshot.

Fixes: https://github.com/w3c/webdriver/issues/1245

This was originally fixed in 3d0755c, but accidentally got partly
reverted in its merge into master in 4265381.

Signed-off-by: Sam Sneddon <gsnedders@apple.com>