Correct handling of Promise values
Previously, Execute Script and Execute Async Script were updated to
invoke the "execute a function body" algorithm by "promise calling" it.
[1] This did not have the intended effect because "execute a function
body" continued to return WebDriver "success" and "error" values,
Additionally, fulfillment of the promise returned by the user-provided
script did not influence completion of the operation.

Refactor "execute a function body" to return an ECMAScript completion
value so that it may be invoked via "promise call." Update both Execute
Script and Execute Async Script to recognize a synchronously-returned
Promise value in accordance with the discussion which motivated the
original change [2].

[1] https://github.com/w3c/webdriver/commit/7307dd659ac75063dea6e0c9261f007faea835df
[2] https://github.com/w3c/webdriver/issues/381