user prompt handler: stop pretending it is an enumerated attribute
Changes the default value of the user prompt handler to "dismiss
and notify", which is effectively what null evaluates to through
the handler's treatment of the missing value default state.

By making the user prompt handler not be an enumerated attribute,
we can remove the "missing value default state" and "not in the
table of simple dialogs" cases when running the handling code.

This change makes the capability behave like other similar capabilities
such as pageLoadStrategy, where the matched/negotiated value is
returned when the session is created.

Fixes: https://github.com/w3c/webdriver/issues/1276