clear algorithm: value sanitization should be done when type attribute is defined
The clear algorithm for <input> uses a list to tell the drivers what
to do.  The last condition, "if the type attribute's current state
defines one" seems like another item to this list but is actually
a logic condition for the preceding sentence, "and then invoke the
value sanitization algorithm".

This turns it into a logical condition using "if-and-only-if".