in-view center point: floor coordinates
Calculating the centre point of a square may produce a decimal
precision number.  Given a 1x1 px square the center point will be
(0.5,0.5), but the DOM operates in integer CSS pixels.

Rounding the number would get us the closest positive integer, or
1.  However, the coordinate (1,1) is beyond the bounding box of the
square.

This means we need to floor the number to ensure we always get a
click point that is within the client rect of the element.

An additional complexity is that not all web platform APIs treat
floating point numbers the same way.  DOMElement.elementsFromPoint,
which we use for getting the paint tree to determine if the web
element is interactable, will floor the number which deems the
element interactable.  However, in Gecko, nsIDOMUtils.sendMouseEvent
ceils the number, causing the actual click point to fall beyond the
square.