Add a endpoint for creating a new window.
This adds an endpoint at /session/<session id>/window/new that opens a
new top-level browsing context. This takes a single parameter "type"
that is used as a hint as to the kind of window to create within the
application. The return value includes the handle of the new window and
the type of the window that was created.