Correct specification of parallel operations
Prior to this patch, parallelism in the Execute Script and Execute Async
Script commands was expressed in terms of ECMAScript promises. This was
ambiguous because in HTML, Promises rely on an environment settings
object [1], and such an object is not generally available in the
WebDriver execution context.

Refactor the command algorithms to instead manage parallel execution
using the generic "queue" primitive from the Infra specification.
Refactor the "execute a function body" algorithm to use Promises, where
following the invocation of "prepare to run a script," an environment
settings object is available.

[1] https://html.spec.whatwg.org/#integration-with-the-javascript-job-queue