internal json clone algorithm: reorder to be possible to run all matches
The way the internal JSON clone algorithm is written it will not
be possible for an implementation to hit all variants.  For example,
we return if the value is an instance of an Object before we check
if it has a callable toJSON property.

Another change in this patch is that every JS object can be serialised
and returned, so it does not make sense to return an error at the
end, since everything in JS is an object.