reorganise chapters
This is a high-level reorganisation of the document's chapters.

Most notably, all element related commands are moved into a single
chapter ("Elements"), with subchapters for Interactability, Retrieval,
State, and Interaction.

Various other chapter names have been shortened, which gives a
simpler look to the table of contents and way of referencing the
specification.

There are no functional changes apart from editorial fixups in this
patch.