fix strings throughout
Strings should be formatted with the quotes outside the <code> tag
so that it is clear to the reader the quote marks are not part of
the string itself, like this:

	"<code>foo</code>"

This patch comes along with a few editorial changes, such as the
removal of unnecessary commas in logical groupings, some table
simplification, and indentation changes as a result of breaking up
the text.  Although overall, there were no larger functional changes
made in this patch.