get rid of legacy RFC2119 style
The WebDriver specification is normative in its descriptions, and no
longer depends on RFC2119, and in particular its capitalised styling
of MUST, SHOULD, RECOMMENDED, MAY, et al.

This patch largely retains the words, but de-capitalises them.  In a
few cases it rewrites the sentence to make better sense in English.