Add a /session/<session id>/print endpoint
This produces a paginated version of the current document as a
base64-encoded PDF file. The featureset is based on what's common in
browsers and largely matches the facility offered by Puppeteer/CDP
except this initial version is missing support for controlling the
headers and footers.

In leiu of detailed defintions much of the output is implemenation
defined; there may be specific cases in which we could do better than
this.