Define non-typeable form controls more exhaustively
The definition used for which form control elements not to use the
actions API for is rather sketchy because it includes inline examples
in a match guard.

This change attempts to formalise the definition of non-typeable
form controls.  It also includes a new example section to guide
implementors, since the definition is rather weak (since we can't
tell UAs how to represent anything.)