Add a hook immediately before a new session is created
The hook allows other specs to get access to the session and the
capabilities immediately before the new session response is sent to
the local end. This allows them to set up any state that is connected
to the session, and update the value of any capabilities they own in
the response.