clarify the bahaviour of beforeunloaded prompts
The wording of the current entry implies to some that promots from
beforeunloaded are always dismissed and thus the promot is in effect
never shown and the user will have no chance to interact with it.

This interpretation would lead it to be impossible to interact with them
and test them - which seems to be the purpose of the webdriver spec, so
this would imply that that interpretation is incorrect.

This clarifies the wording to say that the promot is dismissed on
*subsequent* navigations, thus if a prompt was shown and a navigation
attempt was performed then it would be dismissed, however if the cause
of the alter was a navigation then the promot will be shown so that the
user can interact with it.

fixes #1294