Revert "Add Get Session and Async New Session"
This change to the protocol was made a bit prematurely.
At Sauce Labs we have evaluated the implementation details
after the change and found lots of improvements to be made.
Therefor I propose to revert this and come back when we have
found a better solution to this.

Revert "add unknown session error"

This reverts commit 83153937210ef5e797ee03cb4124b4fcb883f8bb.

Revert "make responses between endpoint and intermediary node conform"

This reverts commit 96790d52d672acb5407ef20ef7597abc30fe9870.

Revert "add get session and async new session"

This reverts commit c83c8f0cef5bf06bcf12708c0e240b1fb574da41.