Allow for multiple WebDriver sessions
With BiDi we want to support multiple sessions, at most one of which
can also enable the HTTP frontend. Therefore we add the concept of
multiple sessions, which can have different flags set according to the
protocols they support (http, bidi, etc.).

This explictly passes the session, parameters and url variables down
into each command. In addition the URL variables are handled as a
infra struct rather than as a custom thing with implied variables.

Instead of having a concept of the "current session" we explicitly
pass the session around where required.