Use "[reference] object" in place of "JSON Serialization".
We have historically called a JS object that represents some remote
end object X in the wire protocol "the JSON serialization of X". But
that terminology is confusing; a JSON serialization is presumably a
string or bytes rather than an object. So instead use "the X object"
or, for objects that represent handles to DOM objects, "the X
reference object" to talk about the object we consturct before
serialization.