Fix Respec errors
The problem seemed to be that Respec expected [[Foo]] internal methods
to be associated with a WebIDL interface, which they aren't because
they're in ECMAScript.

The somewhat hacky workaround is to use the ECMAScript type they're
defined on as the interface name like `Object.[[Put]]` in place of
`[[Put]]`. This doesn't seem ideal, but it at least fixes the errors,
and the intent is still understandable.