Factor global state out of the actions algorithm
The main point of this change is to factor global state out of the
actions processing and dispatch algorithms, so that they can also be
used from BiDi where we don't have the same notion of a current
session, or a specific browsing context.

Along the way it turned out there were a few existing spec-internal
bugs in the way we were using the actions algorithms (e.g. calling
things with the wrong arguments), so this turned into a larger
refactor.

The biggest change is that we now don't make a distinction between an
input source and the state of the input source, which means we don't
have to take care to use the state object instead of the source
itself.

In addition, the list of active input sources is removed in favour of
just having a single input state map with one entry per active source.